+++ scripts/deploy/llm/automodel/deploy_triton.py
@@ -80,18 +80,9 @@ def get_args(argv):
         "--device_map",
         nargs="?",
         choices=["auto", "balanced", "balanced_low_0", "sequential"],
-        default="auto",
+        default=None,
         type=str,
         help="Device mapping strategy for model placement (e.g. 'auto', 'sequential', etc)",
-    )
-    parser.add_argument(
-        "-td",
-        "--torch_dtype",
-        nargs="?",
-        choices=["auto", "bfloat16", "float16", "float32"],
-        default="auto",
-        type=str,
-        help="Torch dtype for the model",
     )
     parser.add_argument(
         "-tpp",
@@ -205,7 +196,6 @@ def hf_deploy(argv):
         task=args.task,
         trust_remote_code=args.trust_remote_code,
         device_map=args.device_map,
-        torch_dtype=args.torch_dtype,
         tp_plan=args.tp_plan,
     )